Use set.isdisjoint instead of creating intersection sets
When only the knowledge of whether two sets intersect is needed and not
what the actual intersection is, negating the result of set.isdisjoint
can be used to more efficiently find whether two sets intersect.

This takes a 1000 piece 100% strictness least merges order puzzle's
generate_early from 10s to 7.5s for me, and produces identical output
with the same seed.