Track the number of added pieces instead of calculating it each time
The number of pieces in the current clusters can be tracked and
incremented every time a piece is added.

This avoids recalculating the more costly operation of counting the
total number of pieces in the current clusters each time.

This takes a 1000 piece 100% strictness least merges order puzzle's
generate_early from 7.5s to 5s for me, and produces identical output
with the same seed.

The unused `remove_piece()` function has been modified, but has not been
tested.