Skip creating a new solution when only the number of merges is needed
This removes some of the work that add_piece was doing that is not
necessary for testing if a piece is suitable to create a new solution
given a specific piece_order.

This change alone reduces takes a 1000 piece 100% strictness least
merges order puzzle's generate_early from 5s to 4.4s for me, and
produces identical output with the same seed.

However, the iteration and summation can now be performed in optimised C
code using `sum` and `map`, further reducing from 4.4s to 2.8s.